initial commit for moving some funcX and other logic from worker to runners.py
